qod_python_sdk/pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "tmode-qod"
version = "1.0.0"
description = """The Quality-On-Demand (QoD) API provides programmable interface for developers and other users (capabilities consumers) to request stable latency or throughput managed by Telco networks without the necessity to have an in-depth knowledge of the 4G/5G system or the overall complexity of the Telecom Systems. """
readme = "README.md"
requires-python = ">=3.7"
classifiers = [
    "python >= 3.7",
    "python",
    "Development Status :: 5 - Production/Stable"
]
dependencies = [
    "requests>=2.31.0",
    "pyjwt==2.8.0",
    "cryptography==42.0.7"
]

[project.urls]
Homepage = "https://devedge.t-mobile.com"
Documentation = "https://devedge.t-mobile.com/documentation-hub"
